Move transform schema tests to transform module
The x-pack transform module has several tests related to schmea
validation. This commit moves those tests from x-pack core into the
module itself. The :test:x-content dependency is then isolated to
transform, so that it does not affect other x-pack modules that do not
need it, and that could conflict with the runtime jackson dependency it
brings in.
