+++ x-pack/plugin/transform/build.gradle
@@ -15,8 +15,15 @@ dependencies {
   testImplementation project(path: xpackModule('analytics'))
   testImplementation project(path: ':modules:aggs-matrix-stats')
   testImplementation project(path: xpackModule('spatial'))
+  testImplementation project(":test:x-content")
 
   internalClusterTestImplementation project(":modules:reindex")
+}
+
+restResources {
+  restApi {
+    include '*'
+  }
 }
 
 addQaCheckDependencies()
