+++ x-pack/plugin/transform/src/test/resources/rest-api-spec/schema/transform.get_transform_stats.schema.json
