+++ x-pack/plugin/transform/src/test/resources/rest-api-spec/schema/transform_checkpoint_stats.schema.json
