+++ x-pack/plugin/transform/src/test/resources/rest-api-spec/schema/transform_checkpointing_info.schema.json
