+++ x-pack/plugin/transform/src/test/resources/rest-api-spec/schema/transform_indexer_stats.schema.json
